clip-filter-backend/requirements.txt
torch>=1.10.2
torchvision>=0.11.3
Pillow>=9.0.1
Flask>=2.1.1
flask-cors>=3.0.10
transformers>=4.12.0
scikit-learn>=0.24.0
